Enable MatchingDeclarationName
