+++ core/ui-library/src/main/kotlin/sk/styk/martin/apkanalyzer/core/uilibrary/components/ChipVariant.kt
@@ -1,0 +1,12 @@
+package sk.styk.martin.apkanalyzer.core.uilibrary.components
+
+import androidx.compose.runtime.Immutable
+
+@Immutable
+sealed interface ChipVariant {
+    data object Default : ChipVariant
+    data object Tonal : ChipVariant
+    data object Positive : ChipVariant
+    data object Warning : ChipVariant
+    data object Negative : ChipVariant
+}
