+++ core/ui-library/src/main/kotlin/sk/styk/martin/apkanalyzer/core/uilibrary/components/IconButtonStyle.kt
@@ -1,0 +1,9 @@
+package sk.styk.martin.apkanalyzer.core.uilibrary.components
+
+enum class IconButtonStyle {
+    Standard,
+    StandardMuted,
+    Filled,
+    Outlined,
+    Highlighted,
+}
