+++ core/ui-library/src/main/kotlin/sk/styk/martin/apkanalyzer/core/uilibrary/lazylist/ListItemPosition.kt
@@ -1,0 +1,3 @@
+package sk.styk.martin.apkanalyzer.core.uilibrary.lazylist
+
+enum class ListItemPosition { Single, First, Middle, Last }
