+++ core/ui-library/src/main/kotlin/sk/styk/martin/apkanalyzer/core/uilibrary/theme/ApkAnalyzerColorPalette.kt
