+++ feature/app-detail/impl/src/main/kotlin/sk/styk/martin/apkanalyzer/feature/appdetail/impl/components/InfoRow.kt
@@ -1,0 +1,10 @@
+package sk.styk.martin.apkanalyzer.feature.appdetail.impl.components
+
+import androidx.compose.runtime.Immutable
+
+@Immutable
+internal data class InfoRow(
+    val label: String,
+    val value: String,
+    val rationale: String,
+)
