+++ feature/apps/impl/src/main/kotlin/sk/styk/martin/apkanalyzer/feature/apps/impl/components/AppDataPermission.kt
@@ -1,0 +1,9 @@
+package sk.styk.martin.apkanalyzer.feature.apps.impl.components
+
+import androidx.compose.runtime.Immutable
+
+@Immutable
+sealed interface AppDataPermission {
+    data object UsageAccess : AppDataPermission
+    data object StorageAccess : AppDataPermission
+}
